+++ BattleNetwork/bnAI.h
@@ -123,7 +123,7 @@ template<typename U, typename ...Args>
  * 
  * If a change state request is made inside of a state, change to that state at end of update
  */
-  void Update(float _elapsed) {
+  void Update(double _elapsed) {
     if (isUpdating) return;
 
     isUpdating = true;
