+++ BattleNetwork/bnAirHockeyCardAction.h
@@ -15,7 +15,7 @@ private:
 public:
   AirHockeyCardAction(Character& owner, int damage);
   ~AirHockeyCardAction();
-  void OnUpdate(float _elapsed);
+  void OnUpdate(double _elapsed);
   void OnAnimationEnd();
   void OnEndAction();
   void OnExecute();
