+++ BattleNetwork/bnAlphaIdleState.h
@@ -7,13 +7,13 @@ class AlphaIdleState : public AIState<AlphaCore>
 {
 private:
   // in seconds
-  float cooldown;
+  double cooldown;
 
 public:
   AlphaIdleState();
   ~AlphaIdleState();
 
   void OnEnter(AlphaCore& a);
-  void OnUpdate(float _elapsed, AlphaCore& a);
+  void OnUpdate(double _elapsed, AlphaCore& a);
   void OnLeave(AlphaCore& a);
 };
