+++ BattleNetwork/bnAlphaRocketState.h
@@ -14,6 +14,6 @@ public:
   ~AlphaRocketState();
 
   void OnEnter(AlphaCore& a);
-  void OnUpdate(float _elapsed, AlphaCore& a);
+  void OnUpdate(double _elapsed, AlphaCore& a);
   void OnLeave(AlphaCore& a);
 };
