+++ BattleNetwork/bnCannonCardAction.cpp
@@ -73,7 +73,7 @@ void CannonCardAction::OnExecute() {
   AddAnimAction(6, onFire);
 }
 
-void CannonCardAction::OnUpdate(float _elapsed)
+void CannonCardAction::OnUpdate(double _elapsed)
 {
   CardAction::OnUpdate(_elapsed);
 }
